# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Cafeteria Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  customer: 'customer',
  admin: 'admin',
} as const;

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface CategoryInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CategoryUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount?: number;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  categoryId: number;
  stock?: number;
  isAvailable?: boolean;
  /** @nullable */
  preparationTime?: number | null;
  /** @nullable */
  calories?: number | null;
  isVegetarian?: boolean;
  isSpicy?: boolean;
  tags?: string[];
}

export interface ProductUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  price?: number;
  /** @nullable */
  imageUrl?: string | null;
  categoryId?: number;
  stock?: number;
  isAvailable?: boolean;
  /** @nullable */
  preparationTime?: number | null;
  /** @nullable */
  calories?: number | null;
  isVegetarian?: boolean;
  isSpicy?: boolean;
  tags?: string[];
}

export interface StockUpdate {
  stock: number;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  stock: number;
  isAvailable: boolean;
  /** @nullable */
  preparationTime?: number | null;
  /** @nullable */
  calories?: number | null;
  isVegetarian?: boolean;
  isSpicy?: boolean;
  tags?: string[];
  createdAt: string;
}

export interface CartItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface CartItemUpdate {
  /** @minimum 0 */
  quantity: number;
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImageUrl?: string | null;
  price: number;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  cash: 'cash',
  card: 'card',
  upi: 'upi',
} as const;

export interface OrderInput {
  /** @nullable */
  tableNumber?: string | null;
  /** @nullable */
  notes?: string | null;
  paymentMethod: OrderInputPaymentMethod;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImageUrl?: string | null;
  price: number;
  quantity: number;
  subtotal: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type OrderPaymentMethod = typeof OrderPaymentMethod[keyof typeof OrderPaymentMethod];


export const OrderPaymentMethod = {
  cash: 'cash',
  card: 'card',
  upi: 'upi',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  status: OrderStatus;
  total: number;
  items: OrderItem[];
  /** @nullable */
  tableNumber?: string | null;
  /** @nullable */
  notes?: string | null;
  paymentMethod: OrderPaymentMethod;
  paymentStatus?: OrderPaymentStatus;
  /** @nullable */
  estimatedReadyTime?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type PaymentInputMethod = typeof PaymentInputMethod[keyof typeof PaymentInputMethod];


export const PaymentInputMethod = {
  cash: 'cash',
  card: 'card',
  upi: 'upi',
} as const;

export interface PaymentInput {
  orderId: number;
  method: PaymentInputMethod;
  /** @nullable */
  upiId?: string | null;
}

export interface PaymentConfirm {
  transactionId: string;
}

export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod];


export const PaymentMethod = {
  cash: 'cash',
  card: 'card',
  upi: 'upi',
} as const;

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface Payment {
  id: number;
  orderId: number;
  method: PaymentMethod;
  status: PaymentStatus;
  amount: number;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  upiId?: string | null;
  createdAt: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DailyRevenue {
  date: string;
  revenue: number;
  orders: number;
}

export interface DashboardStats {
  totalRevenue: number;
  totalOrders: number;
  activeOrders: number;
  totalProducts: number;
  totalUsers: number;
  revenueToday: number;
  ordersToday: number;
  ordersByStatus: StatusCount[];
  revenueByDay?: DailyRevenue[];
}

export interface SalesReport {
  totalRevenue: number;
  totalOrders: number;
  averageOrderValue: number;
  dailyData: DailyRevenue[];
}

export interface TopProduct {
  productId: number;
  productName: string;
  /** @nullable */
  productImageUrl?: string | null;
  totalQuantity: number;
  totalRevenue: number;
  /** @nullable */
  categoryName?: string | null;
}

export interface CategoryRevenue {
  categoryId: number;
  categoryName: string;
  revenue: number;
  orderCount: number;
  percentage?: number;
}

export interface TableMenu {
  tableNumber: string;
  categories: Category[];
  products: Product[];
}

export type ListProductsParams = {
categoryId?: number;
search?: string;
available?: boolean;
};

export type ListOrdersParams = {
status?: ListOrdersStatus;
tableNumber?: string;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type GetSalesReportParams = {
from?: string;
to?: string;
};